fix(download): stream download reports periodic overall progress during copy
feat(update): compute continuous progress

+++ src/Kx/Core/Event/Events.cs
@@ -9,7 +9,7 @@ public record StatusEvent(string Text) : IEvent {
     public DateTime OccurredAt { get; } = DateTime.UtcNow;
 }
 
-public record ProgressEvent(int Percent) : IEvent {
+public record ProgressEvent(float Percent) : IEvent {
     public DateTime OccurredAt { get; } = DateTime.UtcNow;
 };
